requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
imbalanced-learn
flask
streamlit
joblibS